Upgrade to Next 10 and EUI 30 (#18)
This PR upgrades all dependencies, most notably Next.js and EUI.

In so doing, the navigation component from EUI that we used has
been removed, so I've restyled the app to look like the EUI style
guide.

Note as well that I had to set `mainFields` in the Webpack config,
as otherwise it was trying to load from the `lib` dir in EUI instead
of `es`. I assume this is due to internal Next change or their
upgrade to Webpack 5.