Extend from Next's ESLint config (#48)
Since version 11.0.0, Next.js provides an integrated ESLint experience
out of the box. This PR reworks the ESLint configuration to extend
from the Next configuration, which essentially allows us to skip a
lot of configuration.

Note that ESLint is still pinned to <8.0.0, and also running
`yarn lint` prints a complaint about the TypeScript version,
which is harmless.